refactor(mdx): extract shared remark/rehype plugins + components
